pyproject.toml
[project]
name = "biox-macguffins"
version = "1.7.0"
description = "Personal collection of Bioinformatics tools and scripts."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "biopython>=1.86",
    "requests>=2.33.0",
]
